Update fNIRSapp.mlapp
Changed the way the channels are cropped at the end of the acquisition to remove NaNs (default values of the buffer when it hasn't been filled up with data). Now it removes valid samples of channels that received more samples than others so they have the same number of samples than the channels with the less number of samples (above 0). This can only be a problem if some channels receive a significantly smaller number of samples, which does not seem to be the case with the current version of the hardware.