Save function modification
Changes the save function to make the saved data consistent with the SD file used to generate the nSD (if that information is available in the nSD file). This means the naming of the optodes will be different during acquisition and analysis. For analysis, the numbering will be consistent with the probe design in AtlasViewer but inconsistent with the physical optodes. The naming of the physical optodes will be consistent with the visualization in ninjaGUI. Note that this only affects data for ninjaNIRS 2020.