Update
Fully implemented the port communication abstraction and modularized the hardware configuration. Now in theory the fNIRS devices can be completely defined by a subfolder of the 'device_functions' folder and it is unnecessary to change anything on the GUI to make it work, assuming the user knows what they are doing.

(It seems to be working)