Update convertBintoSnirfv3.m
I am now saving information about the missing states in a sidecar file. I also moved the calibration data from the snirf AUX to the sidecar file since it wasn't snirf compliant to save it in the snirf file. This NN22sidecar.mat file should be saved in /sourcedata directory off of the dataset root directory in a subject folder hierarchy that follows the main data folder hierarchy.